refactor(chatroom): split ChatroomPanel god file into components
Extract the 12 sub-components + helpers out of the 1066-line ChatroomPanel.tsx
into their own colocated files, leaving ChatroomPanel as orchestration only
(1066 -> 382 lines). No behavior change: ChatroomPanel.test.tsx (20 cases) and
the full chatroom component suite (58) stay green.

New files:
- chatroomScenarios.ts (DemoScenario + DEMO_SCENARIOS)
- chatroomModerator.ts (profileDisplayName, parseModeratorJson, isModeratorMessage)
- ParticipantBar, ModeratorDecisionBubble, TypingIndicator, CollapsibleMessage
- ChatroomMessageRow (+ ChatroomMessagePresenter), ChatroomMessageList
- MetricsSummaryCard, ChatroomHydratingSkeleton, ChatroomEmptyState
- ChatroomSessionPicker

Addresses the audit's P2 god-file finding.
